mavlink: move hotfix out of two if-statements
ThomasRigi: Here's the modification I came up with to force it to always send mavlink over usb. Probably not the best possible way of doing it, but it works. Basically, I had to force launch_mavlink = true and pull the mavlink stream initialisation out of two if-statements so that it gets executed also when there is no incoming data on the bus.